# public/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Checkin System - OpenAPI 3.0
  version: 1.0.0
servers:
  # Added by API Auto Mocking Plugin
  - url: http://backendcheckin.test/api
    description: SwaggerHub API Auto Mocking
  - url: https://virtserver.swaggerhub.com/minhtanle/CheckinSystem/1.0.0
  - url: http://checkin-v2.local/api
...
paths:
...
  /event/qr-checkin:
    post:
      tags:
        - Event
      summary: Checkin by QR code
      operationId: qrCheckin
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/QrCheckinRequest"
          application/json:
            schema:
              $ref: "#/components/schemas/QrCheckinRequest"
      responses:
        "200":
          description: Successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  status_code:
                    type: integer
                    example: 200
                  message_code:
                    type: string
                    enum:
                      - QR_CODE_INVALID
                      - RESOURCE_NOT_FOUND
                      - CLIENT_HAVE_ALDREADY_CHECKIN
                  data:
                    type: string
                    example: null
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    QrCheckinRequest:
      properties:
        code:
          type: string
          description: "ID for update, if not ID for create"
      required:
        - code
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer